2022/02/17시험

+++ controller/mytest.js
@@ -14,13 +14,28 @@ const getMytest = async (req, res) => {
 	}
 };
 
+const getMytest2 = async (req, res) => {
+	const id = req.query.id;
+	// const isparams = req.params;
+	try {
+		const a = await mytest.findByPk(id);
+		res.send({ msg: "성공했습니다.", data: a });
+	} catch (error) {
+		res.send({
+			msg: "접속실패 error ",
+			data: null,
+		});
+	}
+};
+
 
 const postMytest = async (req, res) => {
+	const test = req.body.test;
 	const box = req.body.box;
 	console.log(box);
 	try {
 		if (box) {
-			const a = await mytest.create({ box: box });
+			const a = await mytest.create({ box: box, test: test });
 			res.send({ msg: "성공했습니다.", data: a });
 		} else {
 			res.send({ msg: "데이터를 안넣어요!" });
@@ -35,10 +50,16 @@ const postMytest = async (req, res) => {
 
 
 const deleteMytest = async (req, res) => {
-	const id = req.params.id;
+	const id = req.query.id;
 	try {
-		const a = await mytest.destroy({ where: { id: id } });
-		res.send({ msg: "성공했습니다.", data: a });
+		const a = await mytest.findByPk(id);
+		if(a) {
+			const b = await mytest.destroy({ where: { id: id } });
+			res.send({ msg: "성공했습니다.", data: b });
+		} else {
+			res.send({ mag: "해당 id존재하지 않음"});
+		}
+
 	} catch (error) {
 		res.statue(200).send({
 			msg: "접속실패 error ",
@@ -91,12 +112,12 @@ const isboxpostTest = async (req, res) => {
 // };
 
 const isboxPatch = async (req, res) =>{
-	const { box } = req.body;
-	const { id } = req.query;
+	const { box, id, test } = req.body;
+	// const { id } = req.query;
 	let a;
 	try {
 		if (id) {
-			a = await mytest.update({box:box},{where:{id : id}});
+			a = await mytest.update({box:box, test:test},{where:{id : id}});
 		}
 		res.send({ msg: "업데이트 완료", a});
 	} catch (error) {
@@ -104,4 +125,4 @@ const isboxPatch = async (req, res) =>{
 	}
 };
 
-module.exports = { getMytest, postMytest, deleteMytest, getOneMytest, isboxpostTest, isboxPatch};
+module.exports = { getMytest, postMytest, deleteMytest, getOneMytest, isboxpostTest, isboxPatch, getMytest2};
